BloodMaze+Abjuration
Added to the bloodmaze level. Added an npc that will only talk to the player if they haven't developed pyromancy. Will give them an abjuration spell that negates one hit.

Changed interactors to not crash when there is more than one provided they have all touched a player already. Still requires player to have interacted with all before, so this will need to be fixed.